[Improvement-3369][api] Code clean and improvement
